+++ pydash/strings.py
@@ -1798,11 +1798,13 @@ def url(*paths, **params):
     return urlunsplit((scheme, netloc, path, query, fragment))
 
 
-def words(text):
+def words(text, pattern=None):
     """Return list of words contained in `text`.
 
     Args:
         text (str): String to split.
+        pattern (str, optional): Custom pattern to split words on. Defaults to
+            ``None``.
 
     Returns:
         list: List of words.
@@ -1811,13 +1813,18 @@ def words(text):
 
         >>> words('a b, c; d-e')
         ['a', 'b', 'c', 'd', 'e']
+        >>> words('fred, barney, & pebbles', '/[^, ]+/g')
+        ['fred', 'barney', '&', 'pebbles']
 
     .. versionadded:: 2.0.0
+
+    .. versionchanged:: 3.2.0
+        Added `pattern` argument.
 
     .. verionchanged:: 3.2.0
         Improved matching for one character words.
     """
-    return js_match(text, RE_WORDS)
+    return js_match(text, pattern or RE_WORDS)
 
 
 #